refactor: Mapeadores agrupados na pasta Mapper
